refactor: move multichain utils to correct file
