+++ src/components/common/NetworkSelector/index.tsx
@@ -39,10 +39,10 @@ import PlusIcon from '@/public/images/common/plus.svg'
 import useAddressBook from '@/hooks/useAddressBook'
 import { CreateSafeOnSpecificChain } from '@/features/multichain/components/CreateSafeOnNewChain'
 import { useGetSafeOverviewQuery } from '@/store/api/gateway'
-import { hasMultiChainAddNetworkFeature } from '@/components/welcome/MyAccounts/utils/multiChainSafe'
 import { InfoOutlined } from '@mui/icons-material'
 import { selectUndeployedSafe } from '@/store/slices'
 import { skipToken } from '@reduxjs/toolkit/query'
+import { hasMultiChainAddNetworkFeature } from '@/features/multichain/utils/utils'
 
 const ChainIndicatorWithFiatBalance = ({
   isSelected,
