+++ src/components/new-safe/create/logic/index.ts
@@ -27,7 +27,7 @@ import { activateReplayedSafe, isPredictedSafeProps } from '@/features/counterfa
 import { getSafeContractDeployment } from '@/services/contracts/deployments'
 import { Safe__factory, Safe_proxy_factory__factory, Safe_to_l2_setup__factory } from '@/types/contracts'
 import { createWeb3 } from '@/hooks/wallets/web3'
-import { hasMultiChainCreationFeatures } from '@/components/welcome/MyAccounts/utils/multiChainSafe'
+import { hasMultiChainCreationFeatures } from '@/features/multichain/utils/utils'
 
 export type SafeCreationProps = {
   owners: string[]
